* 64 bit compability changes.
git-svn-id: https://bloopz.zapto.org/svn/nightfall/projekt/trunk@1079 d57568f3-fe0c-0410-b608-e86116ac1d00

+++ src/aibase.cpp
@@ -653,7 +653,7 @@ namespace Game
 
 		int _LuaAIThread(void* arg)
 		{
-			int i = (int) arg;
+			unsigned long i = (unsigned long)arg; // <~~ 64 bit compability
 			SDL_LockMutex(luaAIWaitMutexes[i]);
 
 			luaAIThreadsRunning[i] = true;
